Fix #32 Replace today date by a date provider
